added apps drag and drop
added apps drag and drop as well as katana-rclick attribute which triggers a method when the attribute is present and element is clicked using a right click like katana-click, however needs design meeting to discuss where to save order. for now changes will only appear per instance but will create a ticket for saving order for next reload.